Created a new class "Globals" which contains some things that should be accessible at every level of an artdaq system. This currently is the "rank" of each process (used as the key for TransferPlugin-based data flow), and a pointer to the MetricManager instance of the application (so that classes such as DataSenderManager, which are constructed by the art thread, can still send metrics). Corresponding changes have been made, removing metricMan_ pointer class members and mpi_rank_ or id_ class members.
Added Transfer time, size and rate metrics to DataSenderManager and DataReceiverManager.

+++ artdaq/DAQrate/DataReceiverManager.hh
@@ -10,6 +10,7 @@
 #include "artdaq-core/Data/Fragment.hh"
 #include "artdaq/TransferPlugins/TransferInterface.hh"
 #include "artdaq/DAQrate/detail/FragCounter.hh"
+#include "artdaq-utilities/Plugins/MetricManager.hh"
 
 namespace artdaq {
   class DataReceiverManager;
@@ -25,9 +26,9 @@ public:
   // source of that fragment as its return value.
   //
   // It is a precondition that a sources_sending() != 0.
-  size_t recvFragment(Fragment & frag, size_t timeout_usec = 0);
+  int recvFragment(Fragment & frag, size_t timeout_usec = 0);
 
-  size_t calcSource();
+  int calcSource();
 
   // How many fragments have been received using this DataReceiverManager object?
   size_t count() const;
@@ -35,13 +36,13 @@ public:
   // How many fragments have been received from a particular destination.
   size_t slotCount(size_t rank) const;
 
-  std::set<size_t> enabled_sources() const { return enabled_sources_; }
+  std::set<int> enabled_sources() const { return enabled_sources_; }
 
 private:
 
-  std::map<size_t, std::unique_ptr<artdaq::TransferInterface>> sources_;
-  std::set<size_t> enabled_sources_;
-  size_t current_source_;
+  std::map<int, std::unique_ptr<artdaq::TransferInterface>> sources_;
+  std::set<int> enabled_sources_;
+  int current_source_;
 
   detail::FragCounter recv_frag_count_; // Number of frags received per source.
 };
